feat: delete user upload (#48)
Adds DELETE /user/uploads/:id to API and wires it up in UI

Given we are going with DELETE /user/uploads/:id I didn't add a delete to the client, as this is not a /car route. Happy to add it as well

All flow tested with local setup from website.