fix!: cid prop on web3file is a string
Previously calling `res.files()` on the response from `client.get` would return an array of Files with a `.cid` property that was a CID instance. This PR makes it a string, to be consistent with the our other usages of CID which are all strings. If folks need to upgrade their string CIDs to instances of CID then the `CID.parse(str)` method is the one to use.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>