feat: storage broker API (#51)
This PR makes changes to allow the storage broker service to interact with the site via GraphQL.

It involves splitting the `Deal` collection out into `BatchEntry`, `Batch` and `Deal` to more closely follow the [schema](https://raw.githubusercontent.com/nftstorage/dagcargo/master/maint/pg_schema_diagram.svg) in the storage broker.

`Content` _will_ be included in multiple batches. This may happen during initial deal creation if the same content is referenced by multiple DAGs, when deals are "renewed" or if we decide to change the format of the batch DAG (and want to "re-do" a bunch of deals in the new format).

`BatchEntry` is _not_ denormalized into `Content`, just incase we are ever given a batch root CID and want to easily know what entries it contains. It also ensures we don't hold write locks on thousands of `Content` records when creating a batch.

**Find new CIDs:**

```gql
query {
  findUploadsCreatedAfter(since: "2021-07-08T15:13:08.318949Z" _cursor: null _size: 1000) {
    data {
      content {
        cid
      }
      user {
        issuer
      }
      created
    }
    # cursor value to be used as _cursor for the next page
    after
  }
}
```

**Create a batch:**

```gql
mutation {
  createBatch(data: {
    batchCid: "BATCHROOTCID"
    pieceCid: "PIECECID"
    entries: [{
      cid: "ENTRYCID"
      dataModelSelector: "Links/0/Links/1/..."
    }]
  }) {
    _id
    cid
    pieceCid
  }
}
```

**Create or update a deal:**

```gql
mutation {
  createOrUpdateDeal(data: {
    batchCid: "BATCHROOTCID" # required (for create, ignored for update)
    miner: "f99"
    network: "mainnet"
    chainDealId: 12345 # required
    activation: "2021-07-08T15:13:08.318949Z"
    renewal: "2022-07-08T15:13:08.318949Z" # (expiry date)
    status: "Active" # required
    statusReason: "It worked!"
  }) {
    _id
    batch {
      cid
    }
    status
  }
}
```

**Updated schema:**

![web3 storage schema(3)](https://user-images.githubusercontent.com/152863/125276207-dc959980-e307-11eb-99fb-3f0c81d61b8f.jpg)