refactor: allow chunked adding of batch entries (#87)
The max payload for Fauna is around 10MB, but even then the query backing this operation can only deal with around 10,000 entries before it throws "An internal server error occurred. Please contact Fauna support.".

This PR splits the `createBatch` call into two: 1) `createBatch` and 2) `addEntries`, so that a batch can be created and then entries can be added to the batch. The `createBatch` method no longer takes a list of entries.

You use it like this now:

```gql
# Create a new batch
mutation {
  createBatch(data: { batchCid: "bafybatch", pieceCid: "baga" }) {
    # returns a Batch object
    _id
  }
}

# Adds new entries to the batch, existing entries are ignored.
# Call it multiple times, call it with the same entries, nothing goes wrong.
mutation {
  addBatchEntries(
    batchCid: "bafybatch",
    entries: [{ cid: "QmFile", dataModelSelector: "Links/0/Hash" }]
  ) {
    # returns a Batch object also
    _id
  }
}
```