feat!: fold in content into user uploads list
"content" is a feature of our internal schema. I don't think it helps users. So it's removed from the list uploads api

Is the plan for pagaination that the caller should grab the `created` property of the last item and pass it in as the `before` prop on the next request? We could lift that out into an envelope and have responses be like

```
{
  data: { ... the payload }
  next: '/user/uploads?before=<iso-data-stamp>
}
```

or similar to make that a litte more friendly... Interested in feedback here. Am going to build out the js client method for it and see how it feels.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>