docs: clarify where to find package READMEs (#295)
When I first tried to test the website, I started by trying to run the commands on this page rather than realizing I needed to dig a few folders down to find the right instructions. This PR clarifies the system and adds direct links to readmes for each package, as well as the separate docs repo.