fix: add pin sync queue collection (#369)
This PR moves the processing of "pending" pin items to a separate collection `PinSyncRequest`. This collection is similar to the `PinRequest` collection but instead of requesting a CID be pinned on another service, it's a request for the status (and potentially other information) to be re-synced with Cluster.

This is a significantly smaller collection to query than the `Pin` collection so should stay performant into the future.