chore: test upload via w3 on cron
The pit canary. A minimally viable smoke test for uploading to web3.storage that can run frequently and alert us if it dies.

In it's current form it will test both staging and production with small files and random files sligtly larger than the local add threshold. I'd like to replace the random files with wikipedia pages or similar, but this will do for now.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>