refactor: esm in worker (#902)
Converts the built worker JS to ESM. Uses miniflare for testing.

This enables using durable objects in the worker so that we can have websockets for w3name. When we get websockets, I can write the [IPNS publisher to listen for updates and publish to the DHT](https://github.com/web3-storage/web3.storage/issues/662). I realised it was going to be a PITA to get Cloudflare expected ESM format (export a default object with a fetch function) to work with regular workers in our tests...I ended up adding miniflare to the tests.

refs https://github.com/web3-storage/web3.storage/issues/659
reolves https://github.com/web3-storage/web3.storage/issues/852