chore: pass ENV to build
We use the ENV value during the api worker build step, so ensure it is set.

TODO: The repitition is annoying. we also have 3 seperate ENV contexts to reason about
- runtime ENV as set from .env in dev and wrangler.toml in prod
- build time env used for SENTRY_ so we can send release info to their api on deploy
- build time replaced global vars for COMMITHASH, VERSION, etc used by runtime sentry reporting and /version endpoint.

we need to simplify that. we could replace the build time replaced global vars with true env vars I think.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>