fix: version endpoint info (#1024)
tweak the esbuild define config to set the git info in the api build.

The `define` api only replaces the exact token, where it finds it seperated by whitespace, with the exact value provided. So the key `VERSION` was not matching on `env.VERSION`, so our envAll function now sets these values on the env for us by pulling them from magic globals that esbuild will set.

I removed the `define` config for ENV. This was not getting applied. Now we pass the ENV to the build step in CI, and miniflare gets it from the .env file.

```console
❯ curl "https://api.web3.storage/version"
{"mode":"rw"}

❯ curl "http://127.0.0.1:8787/version"
{"version":"web3-api@5.1.4-dev+9068490","commit":"9068490aaa4dbe037e8e6aa9eaad706a46f11aed","branch":"main","mode":"rw"}
```

see: https://esbuild.github.io/api/#define

fixes: #1016 

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>