fix: verify s3 backup integrity on upload
add `x-amz-checksum-sha256` header on s3 object put requests.

This is nice as we reuse our existing sha256 hash of the car as an integrity check. AWS will calculate the sha256 of the body it receives and send us a 400 BadDigest error if they dont match, to avoid storing CARs that had bits flipped in transit.

It was necessary to update to the latest version of the aws-sdk as support for sha256 integrity checking was not supported in the older one.

Some debug logging is left in for the s3 client in dev mode, as it was an absolute pain to figure out what it was doing otherwise.

see: #1036

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>