fix: verify CAR block bytes match CID.
Do the bytes match the CID?

For blocks whose multihash was calculated with sha2-256 (the default hashing function) we rehash each block's bytes to check it matches blocks proposed CID multihash.

For other hashing functions, we allow the blocks... The cluster ipfs nodes will drop blocks that are invalid, so such DAGs wont ever be marked as pinned. It's not a good UX but at least we wont say we are storing a full DAG that was sent to us malformed.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>