feat: support ipfs-cluster@1.0
The IPFS Cluster v1 release has much improved perf, and some breaking API changes.

In this PR we upgrade the ipfs-cluster client and local test environment to work with the v1 cluster API.

When this PR is merged to main, the staging API credentials must be updated.

The production release of this PR needs to be coordinated with the upgrade of the ipfs-cluster binaries on the prod cluster.

There will be a second PR to update the pins cron, which should be merged to main once the prod cluster is upgraded.

see: #1184

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>