chore: remove parameters from `createRateLimiter` function
Removing the params, as this function is part of the public API, and shouldn't be configurable.
Create a separate rate limiter for avoiding throttling of the 'adds big files' tests.
The test for the actual rate limiter now takes a full 10 seconds, but going to live with that for now.