chore: release-please for db
this is a nudge to finish off the db migration automation.

as is, this will add release please notes to the PRs and increment the package number.

If we do a major change, we'd need to bump the db dep number in the other projects.

TODO:
- [ ] fork prod db to staging on merge to main
- [ ] auto apply migrations to staging
- [ ] snapshot prod db on release
- [ ] auto apply migrations to prod

see: https://github.com/web3-storage/web3.storage/issues/1482
see: https://github.com/web3-storage/web3.storage/issues/1275

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>