feat: put write to cluster behind a flag
Set `ENABLE_ADD_TO_CLUSTER` to send upload to ipfs-cluster after sending response to user.

The current behavior is to always try to write user uploads to cluster. With this PR we disable that but leave a feature flag to restore it if we need to.

See: https://github.com/web3-storage/web3.storage/issues/1784

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>