feat: check dag status with linkdex
Check the DAG structure of an uploaded CAR using linkdex-api, where LINKDEX_URL is set in the env.

The linkdex-api will run the LinkIndexer in a lambda across all the CARs for a given upload, so it can check the structure across multiple partial CARs where that is the case.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>