feat: update preview summary metadata (#1965)
https://www.notion.so/Update-preview-summary-metadata-568ba710a5e94b27ae70099fed00b405

**This:**

With Web3.Storage you get all the benefits of decentralized storage and
content addressing with the frictionless experience you expect in a
modern storage solution. It’s fast, open and it’s free.

**Should be this:**

With web3.Storage you get all the benefits of decentralized storage and
content addressing with the frictionless experience you expect in a
modern storage solution. It’s fast and it's open.

Co-authored-by: heyjay44 <jennifer.brawn@gmail.com>