fix: fix bug when saving payment method while paymentSettings.storage.subscription=null (#1988)
Previously, the 'earlyAdopter' plan id would be PUT to the api, but the
backend doesn't allow this value as a storage subscription price. The
backend doesn't explicitly model 'earlyAdopter' nor 'plans'. But the
equivalent is to have a `null` storage subscription, and this commit
sends that instead.

This pr also adds type hints that would have prevented this bug from
being introduced, and will hopefully help in the future.