fix: update toucan-js and move to deps
The code was importing the old, unpatched version of `toucan-js`. The patch was in `@web3-storage/toucan-js`.

In the meantime, the original `toucan-js` project as adopted the fix to report error.cause, so we switch back and update our `toucan-js` dep here.

see: https://github.com/robertcepa/toucan-js/releases/tag/v2.7.0

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>