feat: AddPaymentMethodForm shows a more end-user-friendly error message when it encounters an error (#2058)
Motivation:
* provide a more helpful error message in situations like this
https://github.com/web3-storage/web3.storage/issues/1987

How:
* If the AddPaymentForm encounters an error, it doesn't render the error
message (that is probably more intended for a developer to debug than
the end-user), it renders in the html a message that gives the user an
affordance to get in touch with us if the error keeps happening, and
doesn't show any dev details. Dev details still appear in the dev
console to aid debugging.