feat: add synchronized, persistent tab groups (#2172)
I was working on the quickstart stuff for w3up and found myself really
wishing that I had synchronized tab groups, so I can split the
quickstart into broad categories (install, create space, upload, view,
etc) and have the details in tabs for the various runtime envs (CLI,
node, browser).

That would feel super awkward with our current non-syncing tabs, but I
think it could be pretty cool if we get the tabs in sync.

Since I was having a bit of the old docs writer's block, I decided to
just do it and ported over the feature from docusaurus, which is where I
originally cribbed the tabs component in the first place.

There are a few minor changes from their implementation, mostly just
using `@next/router` instead of their thing and using jsdoc for types
instead of proper typescript.

There was one thing I copy/pasted that I didn't fully understand, namely
the `useEvent` utility, which it seems is needed to make a dynamic
callback usable as a dependency in another `useCallback`. It's used to
refresh the tab state from local storage after the initial render. It
seems to work, but I'm just calling it out since it was new to me.

No tests yet, but here's a gif of it doing the thing:


![tab-groups](https://user-images.githubusercontent.com/678715/208783777-63ba932b-3ac3-4ddd-8fb3-19acef819502.gif)