feat: use pickup as `CLUSTER_API_URL`
Update `CLUSTER_API_URL` in wrangler.toml to have the api use pickup in place of cluster for handling pinning service requests.

The prod pickup is currently configured to proxy all requests back over to cluster. The plan is to get pickup in the loop so we can control what fraction of requests it should handle from pickyup config.

Note: Merging this PR will trigger a staging deployment, which will apply the change to the staging config. Merging the release-please PR will trigger a prod deploy and the prod config change will be applied then.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>