feat: recheck pin status on GET /pins/:req-id
When using kubo to `ipfs remote pin add` we get a GET request every second checking for an update to the pin status.

Here we use that as a trigger to ask pickup for a status update. This will be handled by a lambda invocation and a single item db read, and will be pretty low impact.

In return we will be able to inform the user as soon as their pin is complete rather than having to wait up to an hour for an async pin update job to complete.

fixes https://github.com/web3-storage/web3.storage/issues/2213

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>