fix: apply workaround to Simple Analytics blocking navigation to /docs (#2238)
The Simple Analytics event handlers being attached to all of our links
is interpreting the path `/docs` as a link to a download and blocking
the default action of this event. It appears this is because it's
mistaking this path as a link to a .doc file. Explicitly setting the
file extensions to their defaults but removing the `doc` extension seems
to fix this. Probably a bad regex on their end.

See #2237