fix: only recheck pin status if pinning or queued (#2236)
pickup will list a pin as "pinning" while it internally retries after an
error. If fetching the DAG times our, or repeatedly errors for some
other reason, pickup wont keep retrying it. The pin status will either
be queued, pinning, pinned, or failed.

change the logic of the pins cron to only recheck pins that are queued
or pinning.

License: MIT

Signed-off-by: Oli Evans <oli@protocol.ai>