feat: add w3up-launch announcement banner (when env.NEXT_PUBLIC_W3UP_LAUNCH_SUNSET_ANNOUNCEMENT_START is set) (#2319)
Motivation:
* https://github.com/web3-storage/secrets/issues/20

Tactics:
* there is a new env var
`NEXT_PUBLIC_W3UP_LAUNCH_SUNSET_ANNOUNCEMENT_START` that is an ISO8601
datetime of when the announcement should appear. If unset, the
announcement will never appear. My intention is to set it only on the
staging environment for acceptance testing. Once we're happy on staging,
we can enable it on production by setting this var to whatever datetime
we want it to appear.

Scope:
* when enabled at after the appropriate time, the banner should appear
on the following routes
  * /
  * /account
  * /account/payment
  * /docs
* the github actions `website` workflow now uses github environments.
This is because I want to use them to configure the staging environment
to have a particular value for
`NEXT_PUBLIC_W3UP_LAUNCH_SUNSET_ANNOUNCEMENT_START`